Fiz alterações no arquivo carrinho.js / criei o arquivo ItensCarrrinho.js e a pasta carrinho na pasta components
